move settings to own file
