+++ Makefile
@@ -73,7 +73,7 @@ yapf:
 	pipenv run yapf --style .yapf --recursive -i $(MODULE)
 
 
-checks: update-po requirements flake8 pylint
+checks: flake8 pylint
 
 flake8:
 	pipenv run python setup.py flake8
